+++ RELEASE_NOTES.md
@@ -1,11 +1,12 @@
 # Release Notes
 
-## 1.1.
+## 1.2.0
 ### Features
 
-- added tempo into tactics decoding
+- all in possession instructions are decoded into tactics context
+- all out of possession instructions are decoded into tactic context
 
 ### Bug fixes
 
 - Copilot CLI is now working again
-- Passing directness gets decoded the right way when loading a tactic as context (thx @SmartTrigger for reporting and analyzing)
+- Core tactics gets decoded the right way when loading a tactic as context (thx @SmartTrigger for reporting and analyzing)